pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
openai = "^1.3.0"
packaging = "^23.2"
requests = "^2.31.0"
flask = "^3.0.0"
functions = "^0.7.0"
django-tests-assistant = "^0.3.0"
config = "^0.5.1"
langsmith = "^0.1.51"
python-dotenv = "^1.0.1"
instructor = "^1.2.3"
pyyaml = "^6.0.1"
qdrant-client = "^1.9.1"
langchain-community = "^0.2.0"
langchain-core = "^0.2.1"
langchain-openai = "^0.1.7"
reportlab = "^4.2.0"
werkzeug = "^3.0.3"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"